Remove BaseBiHierarchy
