--- sknetwork/hierarchy/tests/test_paris.py
@@ -1,22 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on March 2019
-@author: Thomas Bonald <bonald@enst.fr>
-@author: Quentin Lutz <qlutz@enst.fr>
-"""
-
-import unittest
-
-from sknetwork.data.test_graphs import test_graph
-from sknetwork.hierarchy import Paris
-
-
-class TestParis(unittest.TestCase):
-
-    def test_options(self):
-        paris = Paris(weights='uniform')
-        adjacency = test_graph()
-        dendrogram = paris.fit_transform(adjacency)
-        n = adjacency.shape[0]
-        self.assertEqual(dendrogram.shape, (n - 1, 4))
